Use foss version of pysimplegui
The previous version (4.70.1) seems to cause the Windows version to
crash on startup.  Windows 11 Defender also flags the EXE as having a
virus, although it is likely a false positive.

With this FOSS version, it doesn't crash and the executable is not
flagged.  The resulting binary is larger and takes longer to build.