Fix downloader noclean option
The noclean option seems to have gotten lost with the recent changes.
It was being ignored by the scenery installer causing the installer
to always cleanup the downloaded zip files after an install.

Propagate the noclean flag in the SceneryDownloadWorker so that
the cleanup step can be skipped if indicated.

Also update the noclean state of the persistent dl object when the user
changes the checkbox in the GUI and saves the settings.