Suppress trailing = in config file comments
The configparser stores comments as keys so that they can be written
back out to the config file.  It was setting the value of these comment
keys to the empty string, which causes an equal sign to be appended to
the comment in the config file since an empty string is an actual value.

Change the setting of a comment value from an empty string to None
so that the (ugly) equal signs are no longer written out to the
config file.