fuse: Stop throwing exceptions on flush()
X-Plane doesn't write to any the files in the scenery fuse mounts,
which means that this fuse module doesn't need to define flush(),
sync(), and write(). (write was already not defined before this commit)

Rename the flush() and sync() functions so that fuse does not intercept
them.

This fixes a problem with flush() where flush() was calling the OS
flush() function with a bad handle.  The bad handle caused an
exception and triggered exception processing, which was harmless, but
logged the error in the (debug) log and used up some time in the XP
rendering thread.  The call to flush() was probably triggered by XP
closing .ter and .png files found in the scenery.  It isn't clear
why flush() was being called when the files should have been opened
only for reading, making flushing them unnecessary.

This change saves a few dozen microseconds for each file in the scenery
that XP opens, which isn't much.  But there are a lot of these files
and it adds up.