Remove spurious GET_IMG: FAILED! log message
The (normal mode) log message indicating a GET_IMG: FAILED! isn't
really a failure and is misleading.

This message was printed when we took the retry path in the
process_chunk() function and the second wait timed out.  The download
thread was still running and populated the chunk data after the
process_chunk thread gave up on the image creation because of the
second timeout.  This caused the "not chunk_img and chunk.data"
condition to become true and printed the message.
Basically, a race condition.

The warning message is a bit misleading because it prints out a valid
JPEG header since the chunk.data is actually correct.  It was just
late in coming in.

We could call load_from_memory() at this point since the chunk.data
is probably valid, but might be in the process of being copied.  And
it is probably better not to since we've already waited 2 * maxwait,
and treat it as a normal missing chunk.

Also add the thread ID to log messages in process_chunk() to make
them easier to analyze.